+++ fluxo-kmp-conf/src/main/kotlin/fluxo/conf/dsl/container/impl/KmpConfigurationContainerDslImpl.kt
@@ -17,7 +17,6 @@ import fluxo.conf.dsl.container.impl.target.TargetJvmContainer
 import fluxo.conf.dsl.container.impl.target.TargetLinuxContainer
 import fluxo.conf.dsl.container.impl.target.TargetMingwContainer
 import fluxo.conf.dsl.container.impl.target.TargetWasmContainer
-import fluxo.conf.dsl.container.impl.target.TargetWasmNativeContainer
 import fluxo.conf.impl.kotlin.KOTLIN_1_8_20
 import fluxo.conf.impl.kotlin.KOTLIN_1_9_20
 import fluxo.conf.impl.kotlin.KOTLIN_2_0
@@ -44,8 +43,7 @@ internal class KmpConfigurationContainerDslImpl(
     TargetLinuxContainer.Configure,
     TargetMingwContainer.Configure,
     TargetWasmContainer.Configure,
-    TargetAndroidNativeContainer.Configure,
-    TargetWasmNativeContainer.Configure {
+    TargetAndroidNativeContainer.Configure {
 
     override fun <T : KotlinTargetContainer<KotlinTarget>> onTarget(
         type: Class<T>,
@@ -84,7 +82,7 @@ internal class KmpConfigurationContainerDslImpl(
      * @see org.jetbrains.kotlin.gradle.dsl.KotlinMultiplatformExtension.targetHierarchy
      * @see org.jetbrains.kotlin.gradle.dsl.KotlinTargetHierarchyDsl.default
      */
-    override fun allDefaultTargets() {
+    override fun allDefaultTargets(wasmWasi: Boolean) {
         // KotlinMultiplatformExtension.targetHierarchy
         // https://kotlinlang.org/docs/whatsnew1820.html#new-approach-to-source-set-hierarchy
 
@@ -113,7 +111,7 @@ internal class KmpConfigurationContainerDslImpl(
 
             // WASI target is available since Kotlin 1.9.20.
             // Both WASI and JS can be used together since Kotlin 2.0.
-            if (ENABLE_WASM_WASI && kotlinPluginVersion >= KOTLIN_2_0) {
+            if (wasmWasi && kotlinPluginVersion >= KOTLIN_2_0) {
                 wasmWasi()
             }
         }
@@ -135,5 +133,3 @@ internal class KmpConfigurationContainerDslImpl(
 // https://docs.gradle.org/current/userguide/validation_problems.html#implicit_dependency
 private fun isGradleNotFailingOnImplicitTaskDependencies() =
     GradleVersion.current() < GradleVersion.version("8.0")
-
-private const val ENABLE_WASM_WASI = false