+++ fluxo-kmp-conf/src/main/kotlin/fluxo/conf/dsl/container/impl/KmpTargetCode.kt
@@ -31,13 +31,9 @@ internal enum class KmpTargetCode {
     LINUX_X64,
     LINUX_ARM64,
     LINUX_ARM32_HFP,
-    LINUX_MIPS32,
-    LINUX_MIPSEL32,
 
-    MINGW_X86,
     MINGW_X64,
 
-    IOS_ARM32,
     IOS_ARM64,
     IOS_X64,
     IOS_SIMULATOR_ARM64,
@@ -50,15 +46,13 @@ internal enum class KmpTargetCode {
     WATCHOS_ARM64,
     WATCHOS_DEVICE_ARM64,
     WATCHOS_SIMULATOR_ARM64,
-    WATCHOS_X86,
     WATCHOS_X64,
 
     ANDROID_ARM32,
     ANDROID_ARM64,
     ANDROID_X86,
     ANDROID_X64,
-
-    WASM32;
+    ;
 
 
     internal companion object {
@@ -76,7 +70,7 @@ internal enum class KmpTargetCode {
         internal val COMMON_WASM = arrayOf(WASM_JS, WASM_WASI)
         internal val COMMON_JS = COMMON_WASM + JS
 
-        internal val IOS = arrayOf(IOS_ARM32, IOS_ARM64, IOS_SIMULATOR_ARM64, IOS_X64)
+        internal val IOS = arrayOf(IOS_ARM64, IOS_SIMULATOR_ARM64, IOS_X64)
         internal val MACOS = arrayOf(MACOS_ARM64, MACOS_X64)
         internal val OSX = MACOS
         internal val TVOS = arrayOf(TVOS_ARM64, TVOS_SIMULATOR_ARM64, TVOS_X64)
@@ -86,17 +80,16 @@ internal enum class KmpTargetCode {
             WATCHOS_DEVICE_ARM64,
             WATCHOS_SIMULATOR_ARM64,
             WATCHOS_X64,
-            WATCHOS_X86,
         )
         internal val APPLE = IOS + MACOS + TVOS + WATCHOS
 
         internal val LINUX =
-            arrayOf(LINUX_X64, LINUX_ARM64, LINUX_ARM32_HFP, LINUX_MIPS32, LINUX_MIPSEL32)
-        internal val MINGW = arrayOf(MINGW_X64, MINGW_X86)
+            arrayOf(LINUX_X64, LINUX_ARM64, LINUX_ARM32_HFP)
+        internal val MINGW = arrayOf(MINGW_X64)
         internal val UNIX = APPLE + LINUX
         internal val ANDROID_NATIVE =
             arrayOf(ANDROID_ARM32, ANDROID_ARM64, ANDROID_X64, ANDROID_X86)
-        internal val NATIVE = UNIX + MINGW + ANDROID_NATIVE + WASM32
+        internal val NATIVE = UNIX + MINGW + ANDROID_NATIVE
 
         internal val NON_JVM = COMMON_JS + NATIVE
 
@@ -133,29 +126,23 @@ internal enum class KmpTargetCode {
                         KonanTarget.ANDROID_ARM64 -> ANDROID_ARM64
                         KonanTarget.ANDROID_X64 -> ANDROID_X64
                         KonanTarget.ANDROID_X86 -> ANDROID_X86
-                        KonanTarget.IOS_ARM32 -> IOS_ARM32
                         KonanTarget.IOS_ARM64 -> IOS_ARM64
                         KonanTarget.IOS_SIMULATOR_ARM64 -> IOS_SIMULATOR_ARM64
                         KonanTarget.IOS_X64 -> IOS_X64
                         KonanTarget.LINUX_ARM32_HFP -> LINUX_ARM32_HFP
                         KonanTarget.LINUX_ARM64 -> LINUX_ARM64
-                        KonanTarget.LINUX_MIPS32 -> LINUX_MIPS32
-                        KonanTarget.LINUX_MIPSEL32 -> LINUX_MIPSEL32
                         KonanTarget.LINUX_X64 -> LINUX_X64
                         KonanTarget.MACOS_ARM64 -> MACOS_ARM64
                         KonanTarget.MACOS_X64 -> MACOS_X64
                         KonanTarget.MINGW_X64 -> MINGW_X64
-                        KonanTarget.MINGW_X86 -> MINGW_X86
                         KonanTarget.TVOS_ARM64 -> TVOS_ARM64
                         KonanTarget.TVOS_SIMULATOR_ARM64 -> TVOS_SIMULATOR_ARM64
                         KonanTarget.TVOS_X64 -> TVOS_X64
-                        KonanTarget.WASM32 -> WASM32
                         KonanTarget.WATCHOS_ARM32 -> WATCHOS_ARM32
                         KonanTarget.WATCHOS_ARM64 -> WATCHOS_ARM64
                         KonanTarget.WATCHOS_DEVICE_ARM64 -> WATCHOS_DEVICE_ARM64
                         KonanTarget.WATCHOS_SIMULATOR_ARM64 -> WATCHOS_SIMULATOR_ARM64
                         KonanTarget.WATCHOS_X64 -> WATCHOS_X64
-                        KonanTarget.WATCHOS_X86 -> WATCHOS_X86
 
                         else -> {
                             logger?.w("Unexpected KonanTarget: $konanTarget")
@@ -173,7 +160,6 @@ internal enum class KmpTargetCode {
 
         internal fun fromKotlinFamily(family: Family): Array<KmpTargetCode> {
             return when (family) {
-                Family.WASM -> arrayOf(WASM32)
                 Family.ANDROID -> ANDROID_NATIVE
                 Family.MINGW -> MINGW
                 Family.LINUX -> LINUX
@@ -182,7 +168,7 @@ internal enum class KmpTargetCode {
                 Family.TVOS -> TVOS
                 Family.WATCHOS -> WATCHOS
 
-                Family.ZEPHYR -> arrayOf()
+                else -> arrayOf()
             }
         }
     }