+++ fluxo-kmp-conf/src/main/kotlin/fluxo/conf/dsl/container/impl/target/TargetAppleIosContainer.kt
@@ -5,7 +5,6 @@ package fluxo.conf.dsl.container.impl.target
 import fluxo.conf.dsl.container.impl.ContainerContext
 import fluxo.conf.dsl.container.impl.ContainerHolderAware
 import fluxo.conf.dsl.container.impl.KmpTargetCode
-import fluxo.conf.dsl.container.impl.KmpTargetCode.Companion.DEPRECATED_TARGET_MSG
 import fluxo.conf.dsl.container.impl.KmpTargetCode.IOS_SIMULATOR_ARM64
 import fluxo.conf.dsl.container.impl.KmpTargetContainerImpl
 import fluxo.conf.dsl.container.target.AppleIosTarget
@@ -17,15 +16,10 @@ internal abstract class TargetAppleIosContainer<T : KNT>(
     context: ContainerContext,
     name: String,
 ) : KmpTargetContainerImpl<T>(context, name, APPLE_IOS_SORT_ORDER),
-    KmpTargetContainerImpl.NonJvm.Native.Unix.Apple.Ios<T>,
+    KmpTargetContainerImpl.NonJvm.Native.Nix.Apple.Ios<T>,
     AppleIosTarget<T> {
 
     interface Configure : AppleIosTarget.Configure, ContainerHolderAware {
-
-        @Suppress("DEPRECATION", "OVERRIDE_DEPRECATION")
-        override fun iosArm32(targetName: String, configure: AppleIosTarget<KNT>.() -> Unit) {
-            holder.configure(targetName, ::Arm32, KmpTargetCode.IOS_ARM32, configure)
-        }
 
         override fun iosArm64(targetName: String, configure: AppleIosTarget<KNT>.() -> Unit) {
             holder.configure(targetName, ::Arm64, KmpTargetCode.IOS_ARM64, configure)
@@ -43,14 +37,6 @@ internal abstract class TargetAppleIosContainer<T : KNT>(
         }
     }
 
-
-    @Deprecated(DEPRECATED_TARGET_MSG)
-    class Arm32(context: ContainerContext, targetName: String) :
-        TargetAppleIosContainer<KNT>(context, targetName) {
-
-        @Suppress("DEPRECATION", "DEPRECATION_ERROR", "KotlinRedundantDiagnosticSuppress")
-        override fun KotlinMultiplatformExtension.createTarget() = createTarget(::iosArm32)
-    }
 
     class Arm64(context: ContainerContext, targetName: String) :
         TargetAppleIosContainer<KNT>(context, targetName) {